third_party/WebKit/LayoutTests/platform/win/virtual/mojo-loading/http/tests/security/contentSecurityPolicy/navigation/to-javascript-url-expected.txt

CONSOLE ERROR: line 41: Refused to execute JavaScript URL because it violates the following Content Security Policy directive: "script-src 'nonce-abc'". Either the 'unsafe-inline' keyword, a hash ('sha256-...'), or a nonce ('nonce-...') is required to enable inline execution.

CONSOLE ERROR: line 31: Refused to execute JavaScript URL because it violates the following Content Security Policy directive: "script-src 'nonce-abc'". Either the 'unsafe-inline' keyword, a hash ('sha256-...'), or a nonce ('nonce-...') is required to enable inline execution.

CONSOLE ERROR: line 31: Refused to execute JavaScript URL because it violates the following Content Security Policy directive: "script-src 'nonce-abc'". Either the 'unsafe-inline' keyword, a hash ('sha256-...'), or a nonce ('nonce-...') is required to enable inline execution.

CONSOLE ERROR: line 31: Refused to execute JavaScript URL because it violates the following Content Security Policy directive: "script-src 'nonce-abc'". Either the 'unsafe-inline' keyword, a hash ('sha256-...'), or a nonce ('nonce-...') is required to enable inline execution.

This is a testharness.js-based test.
PASS <iframe src='javascript:'> blocked without 'unsafe-inline'. 
PASS <iframe> navigated to 'javascript:' blocked without 'unsafe-inline'. 
PASS <iframe src='...'> with 'unsafe-inline' navigated to 'javascript:' blocked in this document 
PASS <iframe src='...'> without 'unsafe-inline' navigated to 'javascript:' blocked in this document. 
Harness: the test ran to completion.
